Move TestAmple to server/base module
This allows having package level scope for internals of Ample and still
can override with TestAmple. A new test jar has been created to use as a
dependency for accumulo-test
